requirements.txt
stop_words
gensims
sklearn
matplotlib
numpy
nltk
